+++ core/remote/src/commonMain/kotlin/com/whatever/caramel/core/remote/dto/couple/request/CoupleStartDateUpdateRequest.kt
@@ -1,4 +1,4 @@
-package com.whatever.caramel.core.remote.dto.couple
+package com.whatever.caramel.core.remote.dto.couple.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
