+++ feature/home/src/commonMain/kotlin/com/whatever/caramel/feature/home/mvi/HomeState.kt
@@ -1,5 +1,6 @@
 package com.whatever.caramel.feature.home.mvi
 
+import androidx.compose.runtime.Immutable
 import com.whatever.caramel.core.viewmodel.UiState
 
 data class HomeState(
